requirements.txt
asgiref==3.3.1
coreapi==2.3.3
Django==3.1.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
mysqlclient==2.0.1
PyJWT==1.7.1
PyMsgBox==1.0.7
razorpay==1.2.0
requests==2.11.1
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==1.9.5
sqlparse==0.3.0
virtualenv==16.7.8
virtualenvwrapper-win==1.2.5